tests/submission/W3C/domhtml/HTMLLinkElement02.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>link</title>
<link charset="utf-8" href="/support/html/sample.html" hreflang="en" media="screen" rel="Glossary" type="text/html">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-link-charset" rel="help">
<link rel="author" title="W3C">
</head>
<body>
<p>The charset attribute indicates the character encoding of the linked
    resource.

    Retrieve the charset attribute and examine its value.</p>


<script>
var domtest = async_test("The charset attribute indicates the character encoding of the linked     resource.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vcharset;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("link");
      assert_equals(nodeList.length,3);
testNode = nodeList.item(0);
      vcharset = testNode.charset;

      assert_equals(vcharset,"utf-8");
       
    domtest.done();
}));

</script>
<div id="log"></div>
</body></html>
